AI Note: Add new categories & support large YouTube view counts
Expanded YouTube categories to include pets, sports, politics, and science, updating selection and SQL logic accordingly. Refactored view count handling to use u_int64 throughout C code and float in structs, preventing overflow for high-view videos. Improved network error handling and feedback between backend and frontend. Updated UI: "Begin Game" button is now blue and larger. Enhanced debug output and code clarity.

Personal Note: I expanded the categories because I am thinking of changing the million difficulty to category bans

+++ src/globals.h
@@ -22,7 +22,7 @@ typedef struct fonts {
 
 typedef struct ytNode {
 	struct ytNnode* next;
-	int views;
+	float views;
 	char* filePath;
 	char* sViews;
 	char* subs;
